pyproject.toml
[tool.poetry]
name = "maap-py"
version = "4.0.1a0"
description = "Python client API for interacting with the NASA MAAP API"
repository = "https://github.com/MAAP-Project/maap-py"
authors = ["Jet Propulsion Laboratory <jpl@nasa.gov>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{"include" = "maap"}]
exclude = ["test"]
keywords = ["dataset", "granule", "nasa", "MAAP", "CMR"]
classifiers = [
    "Environment :: Console",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3.5",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
backoff = "^2.2.1"
boto3 = "^1.34.139"
configparser = "^7.0.0"
importlib-resources = "^6.4.0"
mapboxgl = "^0.10.2"
pyyaml = "^6.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["s3"], version = "^1.34.139"}
moto = "^5.0.10"
mypy = "^1.10.1"
pytest = "^8.2.2"
responses = "^0.25.3"
types-requests = "<2.32.0.20240602"
types-pyyaml = "^6.0.12.20240311"
pylint = "^3.2.6"
flake8 = "^7.1.0"
pytest-cov = "^5.0.0"

[tool.coverage.run]
source = ['maap']
relative_files = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
